bot-project/requirements.txt
Pillow==7.2.0
discord.py==1.3.4
PyInstaller==3.6
python-dotenv==0.14.0
requests==2.23.0
git+https://github.com/Rapptz/discord.py
